TCP functionality Base

+++ udp-client.py
@@ -1,8 +1,9 @@
 import socket
 import random as r
-#import os
+import os
 import os.path as path
 import sys
+import threading
 
 
 server_address = '192.168.1.152'
@@ -10,17 +11,39 @@ port = 13000
 BUFFER_SIZE = 1024
 X = (r.randint(1,10000))
 
-name = "peer"+ str(X)
+udp_client_socket = socket.socket(socket.AF_INET, socket.SOCK_DGRAM)
+
+
+
+
+pname = "peer"+ str(X)
+
+def join_Network(name):
+    udp_client_socket.sendto(b"J" + name.encode(), (server_address, port))
+    response, address = udp_client_socket.recvfrom(1024)
+
+    if response[8:] == b"peer joined":
+        print("Peer Joined")
+        return int.from_bytes(response[:8], byteorder='big')
+
+
+
+tcp_port = join_Network(pname)
+tcp_client_socket = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
+server_socket = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
+hostname = socket.gethostname()
+server_socket.bind(((socket.gethostbyname(hostname), tcp_port+1)))
+
+print(socket.gethostbyname(hostname), tcp_port,"1")
 
 def get_file_info(data: bytes) -> (str, int):
     return data[8:].decode(), int.from_bytes(data[:8], byteorder='big')
+
 '''
 def send_file( address: (str, int),filename: str):
     file_size = get_file_size(filename)
     size = (file_size).to_bytes(8, byteorder='big')
     name = filename.encode()
-
-
 
     try:
         message = size + name
@@ -44,24 +67,29 @@ def send_file( address: (str, int),filename: str):
         print(f'An error occurred while sending the file:\n\t{e}')
     finally:
         client_socket.close()
-
-def receive_file(conn_socket: socket, file_name: str, file_size: int):
-    # create a new file to store the received data
-
-    file_name += '.temp'
-    # please do not change the above line!
-    with open(file_name, 'wb') as file:
-        retrieved_size = 0
-        try:
-            while retrieved_size < file_size:
-                chunk = conn_socket.recv(BUFFER_SIZE)
-                retrieved_size += len(chunk)
-                file.write(chunk)
-
-        except OSError as oe:
-            print(oe)
-            os.remove(file_name)
 '''
+'''
+def receive_file(address,file_name: str, file_size: int):
+
+    tcp_client_socket.connect(address)
+    tcp_client_socket.send(b"hello")
+    reply = tcp_client_socket.recv(BUFFER_SIZE)
+    if reply == b"ready":
+        file_name += '.temp'
+
+        with open(file_name, 'wb') as file:
+            retrieved_size = 0
+            try:
+                while retrieved_size < file_size:
+                    chunk = tcp_client_socket.recv(BUFFER_SIZE)
+                    retrieved_size += len(chunk)
+                    file.write(chunk)
+
+            except OSError as oe:
+                print(oe)
+                os.remove(file_name)
+'''
+
 def get_file_size(file_name: str) -> int:
     size = 0
     try:
@@ -71,72 +99,90 @@ def get_file_size(file_name: str) -> int:
         sys.exit(1)
     return size
 
-def request_file():
-    udp_client_socket.sendto(b"R", (server_address, port))
-    response, address = udp_client_socket.recvfrom(1024)
-    if response == b"go ahead":
-        filename = input("Filename: ")
-        udp_client_socket.sendto(bytes(filename.encode()), (server_address, port))
-        response,address = udp_client_socket.recvfrom(1024)
-        tcpport, tcpip = get_file_info(response)
-        print(f'{tcpport}, {tcpip} has the file')
+def tcp_client():
+
+    try:
+        while True:
+
+            print(tcp_port)
+
+            print("Started")
+
+            server_socket.listen(1)
+
+            conn,addr = server_socket.accept()
+
+            conn.sendall(b'Ready')
+            data = conn.recv(1024)
+            if data == b"Cool":
+                print("Awesome")
 
 
-def upload_file():
-    udp_client_socket.sendto(b"U", (server_address, port))
-    response, address = udp_client_socket.recvfrom(1024)
-    if response == b"go ahead":
-        filename = input("Filename: ")
-        size = (get_file_size(filename)).to_bytes(8, byteorder='big')
-        udp_client_socket.sendto(size + bytes(filename.encode()), (server_address, port))
+
+    except KeyboardInterrupt as ki:
+        print("Shutting down...")
+    finally:
+
+        tcp_client_socket.close()
+        server_socket.close()
+
+def udp_client():
+
+    def request_file():
+        udp_client_socket.sendto(b"R", (server_address, port))
         response, address = udp_client_socket.recvfrom(1024)
-        if response == b"uploaded":
-            print("working")
-            return
+        if response == b"go ahead":
+            filename = input("Filename: ")
+            udp_client_socket.sendto(bytes(filename.encode()), (server_address, port))
+            response, address = udp_client_socket.recvfrom(1024)
+            tcp_ip, t_port = get_file_info(response)
+            print(tcp_ip,t_port)
+            tcp_client_socket.connect((tcp_ip,t_port+1))
+            response = tcp_client_socket.recv(1024)
+            if response == b"Ready":
+                tcp_client_socket.send(b"Cool")
 
-def join_Network(name):
+    def upload_file():
+        udp_client_socket.sendto(b"U", (server_address, port))
+        response, address = udp_client_socket.recvfrom(1024)
+        if response == b"go ahead":
+            filename = input("Filename: ")
+            size = (get_file_size(filename)).to_bytes(8, byteorder='big')
+            udp_client_socket.sendto(size + bytes(filename.encode()), (server_address, port))
+            response, address = udp_client_socket.recvfrom(1024)
+            if response == b"uploaded":
+                print("working")
+                return
 
-    udp_client_socket.sendto(b"J" + name.encode(), (server_address, port))
-    response, address = udp_client_socket.recvfrom(1024)
-
-    if response[8:] == b"peer joined":
-        print("Peer Joined")
-        return int.from_bytes(response[:8],byteorder='big')
-
-
-if __name__ == "__main__":
-    udp_client_socket = socket.socket(socket.AF_INET, socket.SOCK_DGRAM)
-    server_socket = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
-    tcp_client_socket = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
-    hostname = socket.gethostname()
 
     try:
         while True:
             message = input("Message Type: ")
-            message_type = bytes(message.encode())
-            '''
-            reply = tcp_client_socket.recv(BUFFER_SIZE)
-
-            if reply == b"request":
-                tcp_client_socket.send(b"ready")
-            '''
-
-            if message == "J":
-                tcp_port = join_Network(name)
-                print(tcp_port)
-                server_socket.bind((socket.gethostbyname(hostname), tcp_port))
 
             if message == "U":
                 upload_file()
 
-            if message == "R":
+            elif message == "R":
                 request_file()
 
-            #print(f'Response from {address[0],address[1]} is "{response.decode()}"')
     except KeyboardInterrupt as ki:
         print("Shutting down...")
     finally:
 
         udp_client_socket.close()
-        tcp_client_socket.close()
-        server_socket.close()
+
+if __name__ == "__main__":
+    try:
+        t1 = threading.Thread(target=udp_client)
+        t2 = threading.Thread(target=tcp_client)
+
+        t1.start()
+        t2.start()
+
+        t1.join()
+        t2.join()
+
+    except KeyboardInterrupt as ki:
+        print("Shutting down...")
+
+
